+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Filter by job profiles from the completed jobs. Refs UIDEXP-328.
 * Need new permission(s) to view Data Export settings in UI. Refs UIDEXP-304.
 * Data export - Add accessibility testing to automated tests. Refs UIDEXP-355.
+* Remove information icon from Settings > Data export > Profiles .Refs UIDEXP-363
 
 ## [6.0.0](https://github.com/folio-org/ui-data-export/tree/v6.0.0) (2023-10-13)
 [Full Changelog](https://github.com/folio-org/ui-data-export/compare/v5.4.1...v6.0.0)